+++ frontend/src/pages/dashboard/index.tsx
@@ -9,7 +9,7 @@ import {
 } from '@mui/material';
 // components
 import Page from '../../components/Page';
-import { useLocation, Link, useRoutes, Navigate, useNavigate } from 'react-router-dom';
+import { useLocation, Link, useRoutes, Navigate } from 'react-router-dom';
 import Iconify from '../../components/Iconify';
 import Profile from './profile';
 import Jios from './jios';
@@ -19,7 +19,6 @@ import Betas from './betas';
 import { listColors } from '../../store/reducers/colors';
 import { listWalls } from '../../store/reducers/walls';
 import MessageBarWithStore from './MessageBarWithStore';
-import useRedirectPath from '../../hooks/useRedirectPath';
 
 interface BottomTab {
   path: string;
@@ -90,17 +89,8 @@ const DashboardRouter = () =>
 export default function Dashboard() {
   const theme = useTheme();
   const dispatch = useDispatch();
-  const navigate = useNavigate();
-  const { redirectPath, clearRedirectPath } = useRedirectPath();
 
-  useEffect(() => {
-    if (redirectPath) {
-      navigate(redirectPath.to, redirectPath.options);
-      clearRedirectPath();
-    }
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
+  // TODO: Check and remove if no one using the redux states
   useEffect(() => {
     dispatch(listGyms());
     dispatch(listColors());
