+++ frontend/src/routes/index.tsx
@@ -1,10 +1,11 @@
 import { Suspense, lazy, ElementType } from 'react';
-import { Navigate, useRoutes, useLocation } from 'react-router-dom';
+import { Navigate, useRoutes, useLocation, useNavigate } from 'react-router-dom';
 import useAutoLogin from 'src/hooks/auth/useAutoLogin';
 // components
 import LoadingScreen from '../components/LoadingScreen';
 import CommonGuard from 'src/components/guards/CommonGuard';
 import NoTelegramUsernamePage from 'src/pages/error/NoTelegramUsernameError';
+import useRedirectPath from '../hooks/useRedirectPath';
 
 // ----------------------------------------------------------------------
 
@@ -22,6 +23,9 @@ const Loadable = (Component: ElementType) => (props: any) => {
 };
 
 export default function Router() {
+  const { redirectPath, clearRedirectPath } = useRedirectPath();
+  const navigate = useNavigate();
+
   useAutoLogin();
 
   return useRoutes([
@@ -46,7 +50,16 @@ export default function Router() {
     {
       path: 'dashboard/*',
       element: (
-        <CommonGuard authenticated onboarded>
+        <CommonGuard
+          authenticated
+          onboarded
+          onSuccess={() => {
+            if (redirectPath) {
+              navigate(redirectPath.to, redirectPath.options);
+              clearRedirectPath();
+            }
+          }}
+        >
           <Dashboard />
         </CommonGuard>
       ),
